done for the most part

+++ app/controllers/reviews_controller.rb
@@ -4,7 +4,8 @@ class ReviewsController < ApplicationController
     # before_action :is_owner?, only: [:update, :destroy]
 
     def index
-        render json: Review.all
+        @review = Review.all
+        render json: @review
     end
 
     def show
@@ -12,7 +13,6 @@ class ReviewsController < ApplicationController
 
         book_reviews = Review.find_by(book_id: params[:book_id])
         render json: book_reviews, status: :ok
-        
     end
 
     def create
@@ -28,28 +28,43 @@ class ReviewsController < ApplicationController
     end
 
     def update
-        @review.update!(review_params)
-        render json: @review, status: :accepted
+        review = Review.find_by(id: params[:review_id])
+
+        if review.present?
+            review.update(update_review_params)
+            render json: review, status: :accepted
+        else
+            render json: {errors: "This review does not exist."}, status: 422
+        end
     end
 
     def destroy
-        review = Review.find_by(id: params[:id])
-        review.destroy
-        head :no_content
+        review = Review.find_by(id: params[:review_id])
+        # debugger
+        if review.present?
+            review.destroy
+            head :no_content
+        end
     end
 
     private
 
     def review_params
-        params.permit(:id, :book_id, :user_id, :review_comment, :rating)
+        params.permit(:id, :book_id, :user_id, :review_comment, :rating, :review_id)
+    end
+
+    def update_review_params
+        params.permit(:review_comment, :rating, :updated_at, :id)
     end
 
     def find_review
-        @review = Review.find_by(id: params[:id])
+        review = Review.find_by(id: params[:review_id])
     end
 
     def is_owner?
-        permitted = @review.user_id == current_user.id
+        review = Review.find_by(id: params[:id])
+        debugger
+        permitted = review == current_user.id
         render json: { errors: {User: "is not authorized" }, status: :forbidden}
     end
 end
